+++ src/lib/components/Page.svelte
@@ -73,7 +73,6 @@
 		display: flex;
 		flex-direction: column;
 		gap: calc(1.5 * var(--spacing));
-		max-width: var(--page-width);
 		margin: auto;
 	}